# Optimize cudnn version softmax